pipeline/src/components/gen_dataset/src/requirements.txt
cryptowatch-sdk
pandas
google-cloud-bigquery
pyarrow
google-cloud-aiplatform